Return `deepSearch()` to extensions
